Modified as object oriented
The element of page and keyframe are classed as Kadenz.Page and
Kadenz.Keyframe, respectively.  They are wrapper of element to get status
and modify the element.  Similarly, the controller methods suck as
next-page and next-step are packed into Kadenz.Controller.  The instance
of Kadenz.Controller is placed in static scope as singleton.  The pages
will be controlled by invoking the method in Kadenz.Controller.