kadenz: specification change in keyframe
Changed specification in keyframe.  Before specification change, the
animation is specified with raw css simply but redundantly :

    <keyframe target="box4" css="width: 96px;
                                 transition-property: width, background-color;
                                 transition-duration: 1s;
                                 background-color:red" />

Now the redundant description have been improved to describe in
attributes as following:

    <keyframe target="box4" duration="1s" width="96px" background-color="red" />

The animation is based on css transition, which is unchanged, but some
description had been reduced.  First of the change is that the duration
is specified in attribute directly.  Other modification is the
transition-property is removed.  The transition-property will be set
from key of attributes on animation.

Another modification is the following.
- Neutral style of the element on page transition
- Modified tutorial to change the explanation about element animation