kadenz.pager: added new module Kadenz.Pager
Added new module Kadenz.Pager to project. Kadenz.Pager is presenter
assist tools.  When number key pressed on the presentation, the pager
will be shown.  The Kadnez.Pager is useful to search page and jump to
specified page.

The pager is at the experimental stage as only jumping with text box.
To complete the pager and to implement more module, it is needed to
review the design of core modules.