kadenz: fixed bug with hash in url
When the page is loaded with page number in hush, the page specified in
hush have became to open the page specified in hush.

TODO: Divide to call a member method of window, and core implementation
The process in window.location is implemented in Kadenz.initPages()
method now.  It is needed to divide the process around window and core
modules.