kadenz.pagenum: added new module kadenz.pagenum
The kadenz.pagenum module is page number of the slide.  The page number
will be added into element which has specified class name or default
class name ("pagenum").

The feature to folio to the page is implemented in kadenz.base and core
module, but is contained bugs.  The above codes have been reconstructed
to Kadenz.pagenum and fixed bugs.