kadenz.io: changed key map
The lazer pointer with a feature to turn a page, sends PageUp/PageDown
signals.  The PageUp and PageDown will be mapped as nextStep() and
prevStep() respectively.  furthermore, Space key are added.

The Enter key is conflict to kadenz.pager, because key event is caught
by kadenz.io and kadenz.pager.  Therefore the mechanism which skips to
conflict of the events.